Shuffle circular imports
